javascript
Finished the event listener for the form,by
adding an if else statement where it's checked
what values the array contains based on which
the sign up button will be enabled or disabled.
Left the age and email inputs alone due to the
html side doing enough to make sure they are
sent with the proper information.